changed routes and renamed Newresult.js to Result.js and old one has been renamed as Oldresult.js
